# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:9000
info:
  contact:
    email: support@dagryn.dev
    name: Dagryn Support
    url: https://github.com/mujhtech/dagryn
  description: Local-first, self-hosted developer workflow orchestrator API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Dagryn API
  version: "1.0"
paths:
...
  /api/v1/webhooks/github:
    post:
      consumes:
      - application/json
      description: Handles incoming GitHub webhooks (push, pull_request, installation
        events) and triggers runs for linked projects
      parameters:
      - description: GitHub event type
        in: header
        name: X-GitHub-Event
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "503":
          description: Service unavailable
          schema:
            type: string
      summary: Handle GitHub webhook
      tags:
      - webhooks
...
securityDefinitions:
  APIKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"